+++ SpringCloudDev/springcloud-service-api/springcloud-system-api/src/main/java/org/springclouddev/system/entity/RoleMenu.java
@@ -36,13 +36,13 @@ private Long id;
 	 * 菜单id
 	 */
 	@ApiModelProperty(value = "菜单id")
-	private Integer menuId;
+	private Long menuId;
 
 	/**
 	 * 角色id
 	 */
 	@ApiModelProperty(value = "角色id")
-	private Integer roleId;
+	private Long roleId;
 
 
 }
